backend/requirements-datasets.txt
# Dataset Generation Requirements
# Install with: pip install -r requirements-datasets.txt

# Core dependencies
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0

# Optional: For advanced synthetic data generation
scikit-image>=0.21.0
scipy>=1.11.0
matplotlib>=3.7.0

# Optional: For ML model training
tensorflow>=2.13.0
scikit-learn>=1.3.0
